# api-spec.yaml
openapi: 3.0.3
info:
  title: StreamVault API
  description: |
    Professional video hosting platform API. Authenticate with a JWT Bearer token or API key (`sv_live_xxx`).
    
    All timestamps are **UNIX epoch integers** (seconds).
  version: 1.0.0
  contact:
    name: StreamVault Support
    url: https://streamvault.link
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.streamvault.link
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      in: header
      name: X-Workspace-Id
      required: true
      schema:
        type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks in workspace
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Webhook list (secret not returned)
    post:
      tags: [Webhooks]
      summary: Create webhook with HMAC-SHA256 signing
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items:
                    type: string
                    enum: [video.ready, video.failed, transcription.complete, video.deleted, '*']
                  default: ['*']
                secret: { type: string, description: Auto-generated if not provided }
      responses:
        '201':
          description: Webhook created (secret shown once)
...
  /api/videos/{videoId}/transcriptions:
    get:
      tags: [Transcriptions]
      summary: List transcriptions for a video
      parameters:
        - in: path
          name: videoId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Transcription list with status
    post:
      tags: [Transcriptions]
      summary: Start Whisper transcription (Pro+ plan required)
      parameters:
        - in: path
          name: videoId
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  enum: [es, en, fr, de, it, pt, ja, zh, ko, ar, ru]
                  default: en
      responses:
        '202':
          description: Transcription job queued
        '403':
          description: Pro plan upgrade required
  /api/videos/{videoId}/transcriptions/{lang}/subtitles.vtt:
    get:
      tags: [Transcriptions]
      summary: Download subtitles as WebVTT
      security: []
      parameters:
        - in: path
          name: videoId
          required: true
          schema: { type: string }
        - in: path
          name: lang
          required: true
          schema: { type: string }
      responses:
        '200':
          description: WebVTT subtitle file
          content:
            text/vtt:
              schema: { type: string }
...